logger is modified: Now: The mechanism to determine that the run is terminated does not rely ONLY on checking the existence of the process with that parent process PID.
Checking the existence of the parent process
required starting a daemon detached from the run that
is affected by killing the parent process. Implemented procedure for
starting a daemon works under Red Hat, but not SUSE.

NOW: in addition to
checking the existence of the parent processes, the logger check
existence of file "AMPS_LOGGER_TERMINATE" that signals termination of
the run. If the file is found, the logger prints the log is printed and
terminates itself.

The logger creates the file "AMPS_LOGGER_TERMINATE" when the code sits
in a  function time that exceeds a predetermined limit. In such a case:
A log is created.
"AMPS_LOGGER_TERMINATE" is created.
Logger waits for 10 sec.
Logger terminates the parent process.
The logger's Server() is called 1/10 sec, so 10 sec of waiting time is
sufficient.

The best part - it works on Pleiades.

+++ src/general/logger.h
@@ -204,6 +204,9 @@ public:
    void InitLogger(int InMpiRank) {
      thread_mpi_rank=InMpiRank; 
 
+     //remove the terminate signal file 
+     remove("AMPS_LOGGER_TERMINATE");
+
     //generate unique file name
     do {
       for (int i=0;i<fname_length-1;i++) {
@@ -358,6 +361,15 @@ public:
              printf("%s\n",msg);
 
              PrintLog(msg);
+
+
+             //output the termination code file and waite a bit before killing the parent process 
+             FILE *fsignal=fopen("AMPS_LOGGER_TERMINATE","w");
+             fprintf(fsignal,"Oops....");
+             fclose(fsignal);
+
+             sleep(10);
+
              kill(data_ptr->parent_pid,9); ////SIGINT);
 
              remove_semaphore();
@@ -368,11 +380,30 @@ public:
 
        sem_post(sem_id);
 
+       //verify that the terminate signal file is present
+       struct stat buffer;
+
+       if (stat("AMPS_LOGGER_TERMINATE",&buffer)==0) {
+         //the file does not exist
+         
+         printf("The parent process (%i) was terminated: output the log\n",data_ptr->parent_pid);
+         PrintLog();
+         remove_semaphore();
+         exit(0);         
+       } 
+
        //verify that the parent process is still alive
        if (0!=kill(data_ptr->parent_pid,0)) {
          //the parent process was exited 
          printf("The parent process (%i) was terminated: output the log\n",data_ptr->parent_pid);
          PrintLog();
+
+         //output the termination code file and waite a bit before killing the parent process 
+         FILE *fsignal=fopen("AMPS_LOGGER_TERMINATE","w");
+         fprintf(fsignal,"Oops....");
+         fclose(fsignal);
+         
+         sleep(10);
 
          remove_semaphore();
          exit(0);
